app/requirements.txt
astroid==2.12.13
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
dill==0.3.6
idna==3.4
influxdb-client==1.35.0
install==1.3.5
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.5.4
pylint==2.15.7
python-dateutil==2.8.2
python-dotenv==0.21.0
reactivex==4.0.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
